Move database initialisation to a junit5 extension so that we can use the junit store.
